implement update Arguments in ScriptModal (Running Pipeline)
