+++ frontend/lost/src/components/pipeline/src/globalComponents/modals/ScriptArgumentsTable.js
@@ -1,16 +1,7 @@
 import React from 'react'
-import IconButton from '../../../../BasicComponents/IconButton'
-import {faCloudUploadAlt} from '@fortawesome/free-solid-svg-icons'
-import pipelineActions from '../../../../../actions/pipeline/pipelineRunning'
-import {useDispatch} from 'react-redux'
 
 
 const ArgumentsTable = (props) => {
-    const dispatch = useDispatch()
-
-    const updateArguments = async () =>{
-        dispatch(pipelineActions.updateArguments())
-    }
     if(props.data){
         return (
             <div style={{ marginLeft: 15, marginRight: 15 }}>
@@ -29,16 +20,6 @@ const ArgumentsTable = (props) => {
                         })}
                     </tbody>
                 </table>
-                {props.showUpdateButton && (
-                <IconButton
-                color="success"
-                icon={faCloudUploadAlt}
-                text="Update Arguments"
-                onClick={updateArguments}
-            />
-
-                )}
-
             </div>
         )
     }else {
